+++ src/langs/en.json
@@ -14,7 +14,7 @@
   "REPLACE_IMAGE": "Replace the image",
   "DRAG_DROP_EXERCISE_TITLE": "Drag and drop these labels into the correct frame of the image",
   "SAVE": "Save",
-  "PREVIEW_NOTE": "Here is what your application will look like from the player's perspective. To modify any settings, please return to the previous steps.",
+  "PREVIEW_NOTE": "Here is what your application will look like from a reader's perspective. To modify any settings, please return to the previous steps.",
   "ADD_LABEL": "Add",
   "UNEXPECTED_ERROR": "Sorry, something went wrong!",
   "START_ADDING_LABEL_HELPER_TEXT": "To add a new label, click on the desired location on the image!"
