+++ src/modules/builder/configuration/AddLabelsStep/ImageFrame.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useRef } from 'react';
+import { useRef } from 'react';
 
 import { Alert, Skeleton, styled } from '@mui/material';
 
@@ -6,8 +6,7 @@ import { Settings, SettingsKeys } from '@/@types';
 import { useAppTranslation } from '@/config/i18n';
 import { hooks } from '@/config/queryClient';
 import { APP } from '@/langs/constants';
-import { useImageDimensionsContext } from '@/modules/context/imageDimensionContext';
-import { debounce } from '@/utils';
+import { useImageObserver } from '@/modules/context/imageDimensionContext';
 
 const Container = styled('div')(() => ({
   display: 'flex',
@@ -20,55 +19,29 @@ const Container = styled('div')(() => ({
 const ImageFrame = (): JSX.Element | null => {
   const { data: appSettings, isLoading: settingLoading } =
     hooks.useAppSettings<Settings>();
+  const imgRef = useRef<HTMLImageElement | null>(null);
 
   const image = appSettings?.find(({ name }) => name === SettingsKeys.File);
   const appSettingId = image?.id || '';
 
   const {
     data: dataFile,
-    isLoading,
+    isLoading: isImageLoading,
     isError,
   } = hooks.useAppSettingFile({
     appSettingId,
   });
 
-  const { imgRef, dimension, saveImageDimension, settingsData } =
-    useImageDimensionsContext();
+  useImageObserver({ imgRef });
 
   const { t } = useAppTranslation();
-
-  const debouncedSaveImageDimension = useRef(
-    debounce(saveImageDimension, 200),
-  ).current;
-
-  useEffect((): (() => void) => {
-    // watch image resize to save image dimension
-    const id = settingsData?.[0]?.id;
-    const onImageSizeChange = (entries: ResizeObserverEntry[]): void => {
-      const entry = entries[0];
-      const { width, height } = entry.contentRect;
-
-      if ((width !== dimension.width || height !== dimension.height) && id) {
-        debouncedSaveImageDimension({ width, height }, id);
-      }
-    };
-
-    const resizeObserver = new ResizeObserver(onImageSizeChange);
-    if (imgRef?.current) {
-      resizeObserver.observe(imgRef.current);
-    }
-
-    return () => {
-      resizeObserver.disconnect();
-    };
-  }, [debouncedSaveImageDimension, dimension, imgRef, settingsData]);
 
   if (dataFile) {
     return (
       <Container>
         <img
           src={URL.createObjectURL(dataFile)}
-          alt="frame"
+          alt="app-frame"
           ref={imgRef}
           style={{
             maxWidth: '100%',
@@ -83,7 +56,7 @@ const ImageFrame = (): JSX.Element | null => {
     );
   }
 
-  if (isLoading || settingLoading) {
+  if (isImageLoading || settingLoading) {
     return <Skeleton />;
   }
