+++ src/modules/context/imageDimensionContext.tsx
@@ -1,46 +1,21 @@
-import React, { createContext, useContext, useMemo, useRef } from 'react';
-
-import { AppSetting } from '@graasp/sdk';
+import React, { useCallback, useEffect, useMemo } from 'react';
 
 import { Settings, SettingsKeys } from '@/@types';
 import { hooks, mutations } from '@/config/queryClient';
-
-const ImageDimensionsContext = createContext<ImageDimensionContextType>({
-  imgRef: null,
-  saveImageDimension: () => {},
-  dimension: { width: 0, height: 0 },
-});
+import { debounce } from '@/utils';
 
 type Props = {
-  children: JSX.Element;
-};
-
-type ImageDimensionContextType = {
-  saveImageDimension: (
-    d: { width: number; height: number },
-    id: string,
-  ) => void;
   imgRef: React.MutableRefObject<HTMLImageElement | null> | null;
-  dimension: { width: number; height: number };
-  settingsData?: AppSetting<Settings>[];
 };
-
-export const ImageDimensionsProvider = ({ children }: Props): JSX.Element => {
-  const imgRef = useRef<HTMLImageElement | null>(null);
+export const useImageObserver = ({ imgRef }: Props): void => {
   const { mutate: patchSetting } = mutations.usePatchAppSetting();
 
   const { data: settingsData } = hooks.useAppSettings<Settings>({
     name: SettingsKeys.SettingsData,
   });
 
-  const value: ImageDimensionContextType = useMemo(() => {
-    const data = settingsData?.[0];
-    const dimension = data?.data?.imageDimension || {
-      width: 0,
-      height: 0,
-    };
-
-    const saveImageDimension = (
+  const saveImageDimension = useCallback(
+    (
       imageDimension: {
         width: number;
         height: number;
@@ -51,7 +26,7 @@ export const ImageDimensionsProvider = ({ children }: Props): JSX.Element => {
         const labels = settingsData?.[0]?.data.labels;
         const prevDimension = settingsData?.[0]?.data.imageDimension;
 
-        // get new offsets in case image dimension change
+        // Calculate new offsets if image dimensions have changed
         const newLabels = labels?.map((label) => {
           const { x, y } = label;
           if (
@@ -74,17 +49,39 @@ export const ImageDimensionsProvider = ({ children }: Props): JSX.Element => {
 
         patchSetting({ id, data: payload });
       }
+    },
+    [settingsData, patchSetting],
+  );
+
+  const debouncedSaveImageDimension = useMemo(
+    () => debounce(saveImageDimension, 200),
+    [saveImageDimension],
+  );
+
+  useEffect((): (() => void) => {
+    const data = settingsData?.[0];
+    const dimension = data?.data?.imageDimension || {
+      width: 0,
+      height: 0,
+    };
+    // watch image resize to save image dimension
+    const id = settingsData?.[0]?.id;
+    const onImageSizeChange = (entries: ResizeObserverEntry[]): void => {
+      const entry = entries[0];
+      const { width, height } = entry.contentRect;
+
+      if ((width !== dimension.width || height !== dimension.height) && id) {
+        debouncedSaveImageDimension({ width, height }, id);
+      }
     };
 
-    return { imgRef, saveImageDimension, dimension, settingsData };
-  }, [imgRef, patchSetting, settingsData]);
+    const resizeObserver = new ResizeObserver(onImageSizeChange);
+    if (imgRef?.current) {
+      resizeObserver.observe(imgRef.current);
+    }
 
-  return (
-    <ImageDimensionsContext.Provider value={value}>
-      {children}
-    </ImageDimensionsContext.Provider>
-  );
+    return () => {
+      resizeObserver.disconnect();
+    };
+  }, [imgRef, debouncedSaveImageDimension, settingsData]);
 };
-
-export const useImageDimensionsContext = (): ImageDimensionContextType =>
-  useContext<ImageDimensionContextType>(ImageDimensionsContext);
