+++ src/utils/index.ts
@@ -1,8 +1,7 @@
 // eslint-disable-next-line @typescript-eslint/ban-types
 export const debounce = (func: Function, delay: number) => {
   let timer: ReturnType<typeof setTimeout>;
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  return (...args: any[]) => {
+  return (...args: unknown[]) => {
     clearTimeout(timer);
     timer = setTimeout(() => func(...args), delay);
   };
